+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Reworked `PdfPngParser` class.
+- Added missing override attribute.
 - Updated internal function of `PdfPngParser` class.
 - Updated `README.md` document.